fix : Calculate의 함수의 매개변수 변경

+++ Software_Rendering/PhongShader.h
@@ -15,5 +15,5 @@ public:
     Vector3 ambient;
     Vector3 reflectDirection;
     PhongShader(Vector3 normalVector, const Vector3& worldPositon1, const Vector3&worldPositon2, const Vector3 &worldPositon3);
-    Vector3 Calculate(Vector3 rgb,std::pair<int,int> UV);
+    Vector3 Calculate(Vector3 rgb,std::pair<float,float> UV);
 };
